refactored pairwise into a unit test
