+++ sdk/src/test/assets/Pairwise.EC.json
