+++ sdk/src/test/java/com/microsoft/did/sdk/crypto/pairwise/TestPairwiseKey.kt
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved
 
-package com.microsoft.did.sdk.crypto.keys
+package com.microsoft.did.sdk.crypto.pairwise
 
 import kotlinx.serialization.Serializable
